feat: modularize Step3D sections
